Move modules from auth::common into auth.
