--- src/daemon/auth/common/mod.rs
@@ -1,13 +1,0 @@
-#[cfg(feature = "multi-user")]
-pub mod crypt;
-
-#[derive(Debug, Clone)]
-pub struct NoResourceType;
-impl std::fmt::Display for NoResourceType {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "<No Resource>")
-    }
-}
-
-#[cfg(feature = "multi-user")]
-pub mod session;
