+++ src/daemon/auth/mod.rs
@@ -1,12 +1,14 @@
-pub mod authorizer;
-pub mod providers;
 
-pub mod common;
 
 pub use self::authorizer::{AuthInfo, Authorizer, LoggedInUser};
 pub use self::permission::{Permission, PermissionSet};
 pub use self::roles::{Role, RoleMap};
 
+pub mod providers;
+
+mod authorizer;
+#[cfg(feature = "multi-user")] mod crypt;
 mod permission;
 mod roles;
+#[cfg(feature = "multi-user")] mod session;
